// src/modules/user/user.enum.ts
export enum UserFieldName {
  ID = "_id",
  EMAIL = "email",
  PASSWORD = "password",
  NAME = "name",
  PHONE = "phone",
  ROLE = "role",
  AVATAR_URL = "avatar_url",
  IS_ACTIVE = "is_active",

  IS_VERIFIED = "is_verified",
  VERIFICATION_TOKEN = "verification_token",
  VERIFICATION_TOKEN_EXPIRES = "verification_token_expires",
  TOKEN_VERSION = "token_version",
  
  LAST_RESET_PASSWORD_AT = "last_reset_password_at",
}